requirements.txt
fastapi
uvicorn
youtube-transcript-api
pydantic
openai-whisper